+++ src/main/java/org/pcap4j/packet/IcmpV4DestinationUnreachablePacket.java
@@ -7,16 +7,22 @@
 
 package org.pcap4j.packet;
 
+import static org.pcap4j.util.ByteArrays.*;
+import java.util.ArrayList;
+import java.util.List;
+import org.pcap4j.util.ByteArrays;
+
 /**
  * @author Kaito Yamada
  * @since pcap4j 0.9.11
  */
-public final class IcmpV4DestinationUnreachablePacket extends IcmpV4UnusedPacket {
+public final class IcmpV4DestinationUnreachablePacket
+extends IcmpV4InvokingPacketPacket {
 
   /**
    *
    */
-  private static final long serialVersionUID = -8841091974928291870L;
+  private static final long serialVersionUID = 2435425895590241470L;
 
   private final IcmpV4DestinationUnreachableHeader header;
 
@@ -26,11 +32,22 @@ public final class IcmpV4DestinationUnreachablePacket extends IcmpV4UnusedPacket
    * @return a new IcmpV4DestinationUnreachablePacket object.
    */
   public static IcmpV4DestinationUnreachablePacket newPacket(byte[] rawData) {
-    return new IcmpV4DestinationUnreachablePacket(rawData);
+    IcmpV4DestinationUnreachableHeader header
+      = new IcmpV4DestinationUnreachableHeader(rawData);
+    byte[] rawPayload
+      = ByteArrays.getSubArray(
+          rawData,
+          header.length(),
+          rawData.length - header.length()
+        );
+    return new IcmpV4DestinationUnreachablePacket(header, rawPayload);
   }
 
-  private IcmpV4DestinationUnreachablePacket(byte[] rawData) {
-    this.header = new IcmpV4DestinationUnreachableHeader(rawData);
+  private IcmpV4DestinationUnreachablePacket(
+    IcmpV4DestinationUnreachableHeader header, byte[] rawData
+  ) {
+    super(rawData);
+    this.header = header;
   }
 
   private IcmpV4DestinationUnreachablePacket(Builder builder) {
@@ -39,17 +56,23 @@ public final class IcmpV4DestinationUnreachablePacket extends IcmpV4UnusedPacket
   }
 
   @Override
-  public IcmpV4DestinationUnreachableHeader getHeader() { return header; }
+  public IcmpV4DestinationUnreachableHeader getHeader() {
+    return header;
+  }
 
   @Override
-  public Builder getBuilder() { return new Builder(this); }
+  public Builder getBuilder() {
+    return new Builder(this);
+  }
 
   /**
    * @author Kaito Yamada
    * @since pcap4j 0.9.11
    */
   public static
-  final class Builder extends org.pcap4j.packet.IcmpV4UnusedPacket.Builder {
+  final class Builder extends org.pcap4j.packet.IcmpV4InvokingPacketPacket.Builder {
+
+    private int unused;
 
     /**
      *
@@ -58,11 +81,22 @@ public final class IcmpV4DestinationUnreachablePacket extends IcmpV4UnusedPacket
 
     private Builder(IcmpV4DestinationUnreachablePacket packet) {
       super(packet);
+      this.unused = packet.getHeader().unused;
+    }
+
+    /**
+     *
+     * @param unused
+     * @return this Builder object for method chaining.
+     */
+    public Builder unused(int unused) {
+      this.unused = unused;
+      return this;
     }
 
     @Override
-    public Builder unused(int unused) {
-      super.unused(unused);
+    public Builder payload(Packet payload) {
+      super.payload(payload);
       return this;
     }
 
@@ -77,15 +111,12 @@ public final class IcmpV4DestinationUnreachablePacket extends IcmpV4UnusedPacket
    * @author Kaito Yamada
    * @since pcap4j 0.9.11
    */
-  public static
-  final class IcmpV4DestinationUnreachableHeader extends IcmpUnusedHeader {
+  public static final class IcmpV4DestinationUnreachableHeader extends AbstractHeader {
 
     /*
      *   0                            15                              31
      *  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
      *  |                             unused                            |
-     *  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-     *  |IPv4 Header + 64 bits of Original Data Datagram(invokingPacket)|
      *  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
      *
      */
@@ -93,19 +124,61 @@ public final class IcmpV4DestinationUnreachablePacket extends IcmpV4UnusedPacket
     /**
      *
      */
-    private static final long serialVersionUID = -1582230511452370852L;
+    private static final long serialVersionUID = 5245447443272345197L;
+
+    private static final int UNUSED_OFFSET
+      = 0;
+    private static final int UNUSED_SIZE
+      = INT_SIZE_IN_BYTES;
+    private static final int ICMPV4_DESTINATION_UNREACHABLE_HEADER_SIZE
+      = UNUSED_OFFSET + UNUSED_SIZE;
+
+    private final int unused;
 
     private IcmpV4DestinationUnreachableHeader(byte[] rawData) {
-      super(rawData);
+      if (rawData.length < ICMPV4_DESTINATION_UNREACHABLE_HEADER_SIZE) {
+        StringBuilder sb = new StringBuilder(80);
+        sb.append("The data is too short to build an ICMPv4 Destination Unreachable Header(")
+          .append(ICMPV4_DESTINATION_UNREACHABLE_HEADER_SIZE)
+          .append(" bytes). data: ")
+          .append(ByteArrays.toHexString(rawData, " "));
+        throw new IllegalRawDataException(sb.toString());
+      }
+
+      this.unused
+        = ByteArrays.getInt(rawData, UNUSED_OFFSET);
     }
 
     private IcmpV4DestinationUnreachableHeader(Builder builder) {
-      super(builder);
+      this.unused = builder.unused;
+    }
+
+    /**
+     *
+     * @return unused
+     */
+    public int getUnused() { return unused; }
+
+    @Override
+    protected List<byte[]> getRawFields() {
+      List<byte[]> rawFields = new ArrayList<byte[]>();
+      rawFields.add(ByteArrays.toByteArray(unused));
+      return rawFields;
     }
 
     @Override
-    protected String getHeaderName() {
-      return "ICMPv4 Destination Unreachable Header";
+    protected String buildString() {
+      StringBuilder sb = new StringBuilder();
+      String ls = System.getProperty("line.separator");
+
+      sb.append("[ICMPv4 Destination Unreachable Header (")
+        .append(length())
+        .append(" bytes)]")
+        .append(ls);
+      sb.append("  Unused: ")
+        .append(unused)
+        .append(ls);
+      return sb.toString();
     }
 
   }