+++ src/main/java/org/pcap4j/packet/IcmpV4RedirectPacket.java
@@ -1,6 +1,6 @@
 /*_##########################################################################
   _##
-  _##  Copyright (C) 2012  Kaito Yamada
+  _##  Copyright (C) 2012-2013  Kaito Yamada
   _##
   _##########################################################################
 */
@@ -11,21 +11,18 @@ import static org.pcap4j.util.ByteArrays.*;
 import java.net.Inet4Address;
 import java.util.ArrayList;
 import java.util.List;
-import org.pcap4j.packet.factory.PacketFactories;
-import org.pcap4j.packet.namednumber.EtherType;
 import org.pcap4j.util.ByteArrays;
-import org.pcap4j.util.IcmpV4Helper;
 
 /**
  * @author Kaito Yamada
  * @since pcap4j 0.9.11
  */
-public final class IcmpV4RedirectPacket extends AbstractPacket {
+public final class IcmpV4RedirectPacket extends IcmpV4InvokingPacketPacket {
 
   /**
    *
    */
-  private static final long serialVersionUID = 6444855719417675899L;
+  private static final long serialVersionUID = 5987521162450318499L;
 
   private final IcmpV4RedirectHeader header;
 
@@ -35,22 +32,28 @@ public final class IcmpV4RedirectPacket extends AbstractPacket {
    * @return a new IcmpV4RedirectPacket object.
    */
   public static IcmpV4RedirectPacket newPacket(byte[] rawData) {
-    return new IcmpV4RedirectPacket(rawData);
+    IcmpV4RedirectHeader header = new IcmpV4RedirectHeader(rawData);
+    byte[] rawPayload
+      = ByteArrays.getSubArray(
+          rawData,
+          header.length(),
+          rawData.length - header.length()
+        );
+    return new IcmpV4RedirectPacket(header, rawPayload);
   }
 
-  private IcmpV4RedirectPacket(byte[] rawData) {
-    this.header = new IcmpV4RedirectHeader(rawData);
+  private IcmpV4RedirectPacket(IcmpV4RedirectHeader header, byte[] rawData) {
+    super(rawData);
+    this.header = header;
   }
 
   private IcmpV4RedirectPacket(Builder builder) {
-    if (
-         builder == null
-      || builder.gatewayInternetAddress == null
-      || builder.invokingPacket == null
-    ) {
+    super(builder);
+
+    if (builder.gatewayInternetAddress == null) {
       StringBuilder sb = new StringBuilder();
-      sb.append("builder: ").append(builder)
-        .append(" builder.ipV4Packet: ").append(builder.invokingPacket);
+      sb.append("builder.gatewayInternetAddress: ")
+        .append(builder.gatewayInternetAddress);
       throw new NullPointerException(sb.toString());
     }
 
@@ -71,10 +74,10 @@ public final class IcmpV4RedirectPacket extends AbstractPacket {
    * @author Kaito Yamada
    * @since pcap4j 0.9.11
    */
-  public static final class Builder extends AbstractBuilder {
+  public static final class Builder
+  extends org.pcap4j.packet.IcmpV4InvokingPacketPacket.Builder {
 
     private Inet4Address gatewayInternetAddress;
-    private Packet invokingPacket;
 
     /**
      *
@@ -82,8 +85,8 @@ public final class IcmpV4RedirectPacket extends AbstractPacket {
     public Builder() {}
 
     private Builder(IcmpV4RedirectPacket packet) {
+      super(packet);
       this.gatewayInternetAddress = packet.header.gatewayInternetAddress;
-      this.invokingPacket = packet.header.invokingPacket;
     }
 
     /**
@@ -96,13 +99,9 @@ public final class IcmpV4RedirectPacket extends AbstractPacket {
       return this;
     }
 
-    /**
-     *
-     * @param invokingPacket
-     * @return this Builder object for method chaining.
-     */
-    public Builder invokingPacket(Packet invokingPacket) {
-      this.invokingPacket = invokingPacket;
+    @Override
+    public Builder payload(Packet payload) {
+      super.payload(payload);
       return this;
     }
 
@@ -124,31 +123,28 @@ public final class IcmpV4RedirectPacket extends AbstractPacket {
      * +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
      * |                 Gateway Internet Address                      |
      * +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-     * |IPv4 Header + 64 bits of Original Data Datagram(invokingPacket)|
-     * +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
      *
      */
 
     /**
      *
      */
-    private static final long serialVersionUID = 5236093878631799797L;
+    private static final long serialVersionUID = -7093717116891501880L;
 
     private static final int GATEWAY_INTERNET_ADDRESS_OFFSET
       = 0;
     private static final int GATEWAY_INTERNET_ADDRESS_SIZE
       = INET4_ADDRESS_SIZE_IN_BYTES;
-    private static final int INVOKING_PACKET_OFFSET
+    private static final int ICMPV4_REDIRECT_HEADER_SIZE
       = GATEWAY_INTERNET_ADDRESS_OFFSET + GATEWAY_INTERNET_ADDRESS_SIZE;
 
     private final Inet4Address gatewayInternetAddress;
-    private final Packet invokingPacket; // Internet Header + 64 bits of Original Data Datagram
 
     private IcmpV4RedirectHeader(byte[] rawData) {
-      if (rawData.length < INVOKING_PACKET_OFFSET) {
+      if (rawData.length < ICMPV4_REDIRECT_HEADER_SIZE) {
         StringBuilder sb = new StringBuilder(80);
         sb.append("The data is too short to build an ICMPv4 Redirect Header(")
-          .append(INVOKING_PACKET_OFFSET)
+          .append(ICMPV4_REDIRECT_HEADER_SIZE)
           .append(" bytes). data: ")
           .append(ByteArrays.toHexString(rawData, " "));
         throw new IllegalRawDataException(sb.toString());
@@ -156,44 +152,10 @@ public final class IcmpV4RedirectPacket extends AbstractPacket {
 
       this.gatewayInternetAddress
         = ByteArrays.getInet4Address(rawData, GATEWAY_INTERNET_ADDRESS_OFFSET);
-
-      Packet p = PacketFactories.getFactory(EtherType.class)
-                   .newPacket(
-                      ByteArrays.getSubArray(
-                        rawData,
-                        INVOKING_PACKET_OFFSET,
-                        rawData.length - GATEWAY_INTERNET_ADDRESS_SIZE
-                      ),
-                      EtherType.IPV4
-                    );
-
-      if (p instanceof IllegalPacket) {
-        this.invokingPacket = p;
-        return;
-      }
-      else if (p.contains(IllegalPacket.class)) {
-        Packet.Builder builder = p.getBuilder();
-        builder.getOuterOf(IllegalPacket.Builder.class)
-          .payloadBuilder(
-             new UnknownPacket.Builder()
-               .rawData(p.get(IllegalPacket.class).getRawData())
-           );
-        for (Packet.Builder b: builder) {
-          if (b instanceof LengthBuilder) {
-            ((LengthBuilder<?>)b).correctLengthAtBuild(false);
-          }
-          if (b instanceof ChecksumBuilder) {
-            ((ChecksumBuilder<?>)b).correctChecksumAtBuild(false);
-          }
-        }
-        p = builder.build();
-      }
-      this.invokingPacket = IcmpV4Helper.makePacketForInvokingPacketField(p);
    }
 
     private IcmpV4RedirectHeader(Builder builder) {
       this.gatewayInternetAddress = builder.gatewayInternetAddress;
-      this.invokingPacket = builder.invokingPacket;
     }
 
     /**
@@ -204,17 +166,10 @@ public final class IcmpV4RedirectPacket extends AbstractPacket {
       return gatewayInternetAddress;
     }
 
-    /**
-     *
-     * @return invokingPacket
-     */
-    public Packet getInvokingPacket() { return invokingPacket; }
-
     @Override
     protected List<byte[]> getRawFields() {
       List<byte[]> rawFields = new ArrayList<byte[]>();
       rawFields.add(ByteArrays.toByteArray(gatewayInternetAddress));
-      rawFields.add(invokingPacket.getRawData());
       return rawFields;
     }
 
@@ -229,10 +184,6 @@ public final class IcmpV4RedirectPacket extends AbstractPacket {
         .append(ls);
       sb.append("  Gateway Internet Address: ")
         .append(gatewayInternetAddress)
-        .append(ls);
-      sb.append("  Invoking Packet: {").append(ls)
-        .append(invokingPacket)
-        .append("}")
         .append(ls);
 
       return sb.toString();