+++ src/main/java/org/pcap4j/packet/IcmpV4TimeExceededPacket.java
@@ -7,16 +7,21 @@
 
 package org.pcap4j.packet;
 
+import static org.pcap4j.util.ByteArrays.*;
+import java.util.ArrayList;
+import java.util.List;
+import org.pcap4j.util.ByteArrays;
+
 /**
  * @author Kaito Yamada
  * @since pcap4j 0.9.11
  */
-public final class IcmpV4TimeExceededPacket extends IcmpV4UnusedPacket {
+public final class IcmpV4TimeExceededPacket extends IcmpV4InvokingPacketPacket {
 
   /**
    *
    */
-  private static final long serialVersionUID = 4186314279255762737L;
+  private static final long serialVersionUID = -7403391242412189831L;
 
   private final IcmpV4TimeExceededHeader header;
 
@@ -26,11 +31,19 @@ public final class IcmpV4TimeExceededPacket extends IcmpV4UnusedPacket {
    * @return a new IcmpV4TimeExceededPacket object.
    */
   public static IcmpV4TimeExceededPacket newPacket(byte[] rawData) {
-    return new IcmpV4TimeExceededPacket(rawData);
+    IcmpV4TimeExceededHeader header = new IcmpV4TimeExceededHeader(rawData);
+    byte[] rawPayload
+      = ByteArrays.getSubArray(
+          rawData,
+          header.length(),
+          rawData.length - header.length()
+        );
+    return new IcmpV4TimeExceededPacket(header, rawPayload);
   }
 
-  private IcmpV4TimeExceededPacket(byte[] rawData) {
-    this.header = new IcmpV4TimeExceededHeader(rawData);
+  private IcmpV4TimeExceededPacket(IcmpV4TimeExceededHeader header, byte[] rawData) {
+    super(rawData);
+    this.header = header;
   }
 
   private IcmpV4TimeExceededPacket(Builder builder) {
@@ -53,7 +66,9 @@ public final class IcmpV4TimeExceededPacket extends IcmpV4UnusedPacket {
    * @since pcap4j 0.9.11
    */
   public static
-  final class Builder extends org.pcap4j.packet.IcmpV4UnusedPacket.Builder {
+  final class Builder extends org.pcap4j.packet.IcmpV4InvokingPacketPacket.Builder {
+
+    private int unused;
 
     /**
      *
@@ -62,11 +77,22 @@ public final class IcmpV4TimeExceededPacket extends IcmpV4UnusedPacket {
 
     private Builder(IcmpV4TimeExceededPacket packet) {
       super(packet);
+      this.unused = packet.getHeader().unused;
+    }
+
+    /**
+     *
+     * @param unused
+     * @return this Builder object for method chaining.
+     */
+    public Builder unused(int unused) {
+      this.unused = unused;
+      return this;
     }
 
     @Override
-    public Builder unused(int unused) {
-      super.unused(unused);
+    public Builder payload(Packet payload) {
+      super.payload(payload);
       return this;
     }
 
@@ -81,15 +107,12 @@ public final class IcmpV4TimeExceededPacket extends IcmpV4UnusedPacket {
    * @author Kaito Yamada
    * @since pcap4j 0.9.11
    */
-  public static
-  final class IcmpV4TimeExceededHeader extends IcmpUnusedHeader {
+  public static final class IcmpV4TimeExceededHeader extends AbstractHeader {
 
     /*
      *   0                            15                              31
      *  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
      *  |                             unused                            |
-     *  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
-     *  |IPv4 Header + 64 bits of Original Data Datagram(invokingPacket)|
      *  +-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+-+
      *
      */
@@ -97,19 +120,61 @@ public final class IcmpV4TimeExceededPacket extends IcmpV4UnusedPacket {
     /**
      *
      */
-    private static final long serialVersionUID = 4981354953509735577L;
+    private static final long serialVersionUID = -4958423484698708497L;
+
+    private static final int UNUSED_OFFSET
+      = 0;
+    private static final int UNUSED_SIZE
+      = INT_SIZE_IN_BYTES;
+    private static final int ICMPV4_TIME_EXCEEDED_HEADER_SIZE
+      = UNUSED_OFFSET + UNUSED_SIZE;
+
+    private final int unused;
 
     private IcmpV4TimeExceededHeader(byte[] rawData) {
-      super(rawData);
+      if (rawData.length < ICMPV4_TIME_EXCEEDED_HEADER_SIZE) {
+        StringBuilder sb = new StringBuilder(80);
+        sb.append("The data is too short to build an ICMPv4 Time Exceeded Header(")
+          .append(ICMPV4_TIME_EXCEEDED_HEADER_SIZE)
+          .append(" bytes). data: ")
+          .append(ByteArrays.toHexString(rawData, " "));
+        throw new IllegalRawDataException(sb.toString());
+      }
+
+      this.unused
+        = ByteArrays.getInt(rawData, UNUSED_OFFSET);
     }
 
     private IcmpV4TimeExceededHeader(Builder builder) {
-      super(builder);
+      this.unused = builder.unused;
+    }
+
+    /**
+     *
+     * @return unused
+     */
+    public int getUnused() { return unused; }
+
+    @Override
+    protected List<byte[]> getRawFields() {
+      List<byte[]> rawFields = new ArrayList<byte[]>();
+      rawFields.add(ByteArrays.toByteArray(unused));
+      return rawFields;
     }
 
     @Override
-    protected String getHeaderName() {
-      return "ICMPv4 Time Exceeded Header";
+    protected String buildString() {
+      StringBuilder sb = new StringBuilder();
+      String ls = System.getProperty("line.separator");
+
+      sb.append("[ICMPv4 Time Exceeded Header (")
+        .append(length())
+        .append(" bytes)]")
+        .append(ls);
+      sb.append("  Unused: ")
+        .append(unused)
+        .append(ls);
+      return sb.toString();
     }
 
   }